VmaUsageを外部から持ってきたらエラーが減った。 次回、assimpライブラリも何か使い方が間違っているようなので修正する

+++ MyVulkan/MyVulkan/RenderScene.cpp
@@ -21,7 +21,7 @@ void RenderScene::Initialize(VmaAllocator* allocator)
 
 void RenderScene::Update()
 {
-	float now = glfwGetTime();
+	float now = (float)glfwGetTime();
 	deltaTime = now - lastTime;
 	lastTime = now;
